combine delete job script with job details
