+++ walk-in-interview/src/main/webapp/job-details/index.html
@@ -52,6 +52,5 @@
 </body>
 
 <script type="module" src="./script.js"></script>
-<script type="module" src="../delete-job/script.js"></script>
 
 </html>
